brain: weigh shorter matches more fairly when thinking
The old algorithm was to weigh matches by the square of the number of matching tokens in the prefix. The idea behind this was that the bot would prefer longer matches as its knowledge grows. However, this goal was already accomplished by the soft limit on the number of options chosen. Using quadratic weight with an order of 6 ended up making the bot almost always repeat messages instead of combining different ones.