brain/sqlbrain: preserve iteration across context drops
When we drop context, we model the result as appending the new search
to the existing sequence, including preserving the skip sequence.
This way, we remain uniform across the entire chain space.

This change also fixes a bug where we were using the same skip sequence
after dropping context, which would lead to almost always not selecting
a term and thereby terminating generation early.